Homepage structure and short CSS style changes
I added new html document and changed it to homepage.html.

For the style I added style.css file so new css changes could be applied to new html file.

This is just a basic structure of future CEE homepage and doesn't contain any detailed styles because of possible future changes that will be discussed later in the process of making the overall web.