Fixed header with drop-down list
The header blurs when the size gets too big since the quality of the photo drops as the screen gets wider. Drop-down list has changed from white to purple colour because it looks better due to blurred background (which wasn't planned when the design was being created but with this I propose that to be changed).